fixed few errors

+++ base/views.py
@@ -1,13 +1,14 @@
 from django.shortcuts import render, redirect
 from .models import Room,Topic
+from django.http import HttpResponse
 from django.db.models import Q
 from .forms import RoomForm, RegistrationForm, LoginForm
-from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth import login, logout
 from django.contrib.auth.decorators import login_required
 from django.contrib import messages
 
 # Create your views here.
-@login_required
+@login_required(login_url='login')
 def home(req):
     q = req.GET.get('q') if req.GET.get('q') != None else ''
 
@@ -20,13 +21,13 @@ def home(req):
     context = {'rooms':rooms, 'topics':topics, 'rooms_count':rooms_count}
     return render(req, 'home.html', context)
 
-@login_required
+@login_required(login_url='login')
 def room(req,pk):
     room = Room.objects.get(id=pk)
     context = {'room':room}
     return render(req, 'room.html',context)
 
-@login_required
+@login_required(login_url='login')
 def createRoom(req):
     form = RoomForm()
     if req.method == 'POST':
@@ -36,10 +37,12 @@ def createRoom(req):
             return redirect('home')
     return render(req, 'room_form.html',{'form':form})
 
-@login_required
+@login_required(login_url='login')
 def updateRoom(req,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    if req.user != room.host:
+        return HttpResponse("You are not allowed update this room!!")
     if req.method =='POST':
         form = RoomForm(req.POST, instance = room)
         if form.is_valid():
@@ -47,9 +50,11 @@ def updateRoom(req,pk):
             return redirect('home')
     return render(req, 'room_form.html',{'form':form})
 
-@login_required
+@login_required(login_url='login')
 def deleteRoom(req,pk):
     room = Room.objects.get(id=pk)
+    if req.user!= room.host:
+        return HttpResponse("You are not allowed to delete this room!!")
     if req.method == 'POST':
         room.delete()
         return redirect('home')
@@ -68,6 +73,9 @@ def registerView(req):
     return render(req, 'register.html',{'form':form})
 
 def loginView(req):
+    if req.user.is_authenticated:
+        return redirect('home')
+    
     form = LoginForm(data=req.POST or None)
     if req.method == "POST":
         if form.is_valid():
@@ -82,5 +90,4 @@ def loginView(req):
 @login_required
 def logoutView(req):
     logout(req)
-    messages.info(req, "You have successfully logged out.") 
     return redirect('login')
